// src/enums/asset-paths.enum.ts
export enum AssetPaths {
    RESUME_IT = "Resume-sondos-thouir-tn.pdf",
    Home_SVG = "assets/illustrations/Home.webp",
    FULL_STACK_DEVELOPMENT_SVG = "assets/illustrations/dev.webp",
    network_SVG = "assets/illustrations/network.webp",
    mobile_SVG = "assets/illustrations/mobile.webp",
    PROJECTS_COVER = "assets/illustrations/ProjectsCover.svg",
    PROJECT_AI = "assets/projects-images/AI.png",
    PROJECT_FIRE_FIGHT= "assets/projects-images/FireFight.jpg",
    PROJECT_INTRANET = "assets/projects-images/Intranet.png",
    PROJECT_ECOMMERCE = "assets/projects-images/ECommerce.png",
    PROJECT_PORTAL = "assets/projects-images/Portal.png",
    PROJECT_DONATION = "assets/projects-images/FloraDon.jpg",
    PROJECT_COVID_ANDROID = "assets/projects-images/fightcovidandroid.png",
    PROJECT_COVID_WEB = "assets/projects-images/fightcovidweb.png",
    PROJECT_INFRASTRUCTURE = "assets/projects-images/Infrastructure.png",
    PROJECT_MPLS = "assets/projects-images/Mpls.png",
    PROJECT_STUDY_MANAGEMENT= "assets/projects-images/insc.png",
    EXPERIENCE_COVER = "assets/illustrations/ExperienceCover.webp",
    EDUCATION_COVER = "assets/illustrations/EducationCover.svg",
    ACHIEVEMENT_COVER = "assets/illustrations/AchievementCover.webp",
    ACHIEVEMENT_CCNA = "assets/achievement-images/CCNA.png",
    ACHIEVEMENT_CCNA2= "assets/achievement-images/CCNA2.jpg",
}